pi/requirements.txt
opencv-python==4.8.1.78
numpy==1.26.4
scipy
PyYAML
requests
pyserial==3.5
pynmea2==1.19.0
paho-mqtt==1.6.1
